Rework identification of OSC sequences.
I didn't like the previous patch setting a flag claiming that two
kinds of thing were APC which aren't in fact APC. So I've made a
little enum to distinguish them in the code. There's an outside chance
we might want to handle some case of these in future, in which case
this makes it easier, but more likely it's just making me feel less
wrong about it.

But I've also folded the two existing flags osc_is_apc and osc_w into
that single enum field, which I think is an improvement.