GTK: tell the input method context about focus changes
GtkIMContext has focus_in and focus_out methods for telling it when the
corresponding widget gains or loses keyboard focus.  It's not obvious to
me why these are necessary, but PuTTY now calls them when it sees
focus-in and focus-out events for the terminal window.  Somehow, this
has caused Hangul input to start working in PuTTY. I can't yet
see what I'm typing for lack of proper preedit support, though.