Very crude support for displaying pre-edit strings
We simply pass each character to term_display_graphic_char and then
put the cursor back where we found it.  This works in simple cases,
but is fundamentally wrong.  Really we should do this in a way that
doesn't touch the terminal state and just gets rendered on top of it
somehow.