Minimal viable pre-edit support
Now we can cope with a single wide or narrow pre-edit character, which
is good enough for the input methods that I use.  When rendering the
line that contains the cursor we set up a little array of termchars
that contains the pre-edit text and work out where it should be
displayed.  Then when rendering the screen we switch between
displaying text from the real terminal and from the pre-edit string as
necessary.

Ideally, we should support longer strings, combining characters, and
setting attributes.  I think the current architecture should make all
of those possible, but not entirely easy.